Docs 20 apr (#2556)
* doc tweaks
* add postgresql in requirements-output.txt
* add ruff fixes for new ruff
* disable pyre and pytype for 3.14-dev

+++ src/cowrie/core/utils.py
@@ -17,7 +17,7 @@ def durationHuman(duration: float) -> str:
     """
     Turn number of seconds into human readable string
     """
-    seconds: int = int(round(duration))
+    seconds: int = round(duration)
     minutes: int
     minutes, seconds = divmod(seconds, 60)
     hours: int
